added query-string helper to reduce code-duplication. also added icon to something else tile
