+++ HelpMyStreetFE/HelpMyStreetFE/Views/RequestHelp/RequestStage/_Activties.cshtml
@@ -13,8 +13,8 @@
         new RequestActivityViewModel{ Activity = HelpMyStreet.Utils.Enums.SupportActivities.PhoneCalls_Friendly,FriendlyName ="Friendly Chat", Description = "A freindly chat on the phone", ColourCode = "light-purple", Icon = "friendlychat.svg"},
         new RequestActivityViewModel{ Activity = HelpMyStreet.Utils.Enums.SupportActivities.PhoneCalls_Anxious,FriendlyName ="Supportive Chat", Description = "Someone to talk through some challenges over the phone", ColourCode = "light-purple", Icon = "supportivechat.svg"},
         new RequestActivityViewModel{ Activity = HelpMyStreet.Utils.Enums.SupportActivities.HomeworkSupport,FriendlyName ="Homework", Description = "Support for children being homeschooled", ColourCode = "dark-blue", Icon = "homework.svg"},
-        
-        new RequestActivityViewModel{ Activity = HelpMyStreet.Utils.Enums.SupportActivities.Other, FriendlyName = "Something Else", Description= "'Please tell us more below", ColourCode = "dark-blue"},
+
+        new RequestActivityViewModel{ Activity = HelpMyStreet.Utils.Enums.SupportActivities.Other, FriendlyName = "Something Else", Description= "Please tell us more below", ColourCode = "dark-blue", Icon="question-mark.svg" }
     };
 }
