+++ HelpMyStreetFE/HelpMyStreetFE/src/js/authenticate.js
@@ -1,14 +1,5 @@
 import { processYoti } from "./profile/yoti";
-
-var getParameterByName = function (name, url) {
-    if (!url) url = window.location.href;
-    name = name.replace(/[\[\]]/g, '\\$&');
-    var regex = new RegExp('[?&]' + name + '(=([^&#]*)|&|#|$)'),
-        results = regex.exec(url);
-    if (!results) return null;
-    if (!results[2]) return '';
-    return decodeURIComponent(results[2].replace(/\+/g, ' '));
-}
+import { getParameterByName } from "./shared/querystring-helper"
 
 $(document).ready(function () {
     var token = getParameterByName("token");
