+++ HelpMyStreetFE/HelpMyStreetFE/src/js/profile/verification.js
@@ -1,4 +1,5 @@
 ﻿import { initialiseYoti } from "./yoti";
+import { updateQueryStringParam, getParameterByName } from "../shared/querystring-helper";
 
 export function initialiseVerification() {
 
@@ -126,37 +127,4 @@ function _setActive(stepNumber) {
 }
 
 
-var updateQueryStringParam = function (key, value) {
-    var baseUrl = [location.protocol, '//', location.host, location.pathname].join(''),
-        urlQueryString = document.location.search,
-        newParam = key + '=' + value,
-        params = '?' + newParam;
-    baseUrl = removeTrailingSlashes(baseUrl)
-    // If the "search" string exists, then build params from it
-    if (urlQueryString) {
-        var keyRegex = new RegExp('([\?&])' + key + '[^&]*');
-
-        // If param exists already, update it
-        if (urlQueryString.match(keyRegex) !== null) {
-            params = urlQueryString.replace(keyRegex, "$1" + newParam);
-        } else { // Otherwise, add it to end of query string
-            params = urlQueryString + '&' + newParam;
-        }
-    }
-    window.history.replaceState({}, "", baseUrl + params);
-};
-
-function removeTrailingSlashes(url) {
-    return url.replace(/\/+$/, ''); //Removes one or more trailing slashes from URL
-}
-
-var getParameterByName = function (name, url) {
-    if (!url) url = window.location.href;
-    name = name.replace(/[\[\]]/g, '\\$&');
-    var regex = new RegExp('[?&]' + name + '(=([^&#]*)|&|#|$)'),
-        results = regex.exec(url);
-    if (!results) return null;
-    if (!results[2]) return '';
-    return decodeURIComponent(results[2].replace(/\+/g, ' '));
-}
