+++ HelpMyStreetFE/HelpMyStreetFE/src/js/profile/yoti.js
@@ -1,3 +1,5 @@
+
+
 export async function processYoti(thisToken, userId, mobile) {
     $('#overlay').show();
     $('.loading-overlay').show();
